+++ src/lib/cache_disk.c
@@ -33,6 +33,7 @@ static int crop_dir(sds cache_basedir, const char *type, int keepdays);
  */
 void cache_disk_clear(struct t_config *config) {
     crop_dir(config->cachedir, DIR_CACHE_COVER, 0);
+    crop_dir(config->cachedir, DIR_CACHE_HTTP, 0);
     crop_dir(config->cachedir, DIR_CACHE_LYRICS, 0);
     crop_dir(config->cachedir, DIR_CACHE_THUMBS, 0);
     crop_dir(config->cachedir, DIR_CACHE_MISC, 0);
@@ -51,6 +52,9 @@ void cache_disk_crop(struct t_config *config) {
     }
     if (config->cache_thumbs_keep_days > CACHE_DISK_DISABLED) {
         crop_dir(config->cachedir, DIR_CACHE_THUMBS, config->cache_thumbs_keep_days);
+    }
+    if (config->cache_http_keep_days > CACHE_DISK_DISABLED) {
+        crop_dir(config->cachedir, DIR_CACHE_HTTP, config->cache_http_keep_days);
     }
     crop_dir(config->cachedir, DIR_CACHE_MISC, config->cache_misc_keep_days);
 }