+++ src/lib/config_def.h
@@ -53,6 +53,7 @@ struct t_config {
     int cache_lyrics_keep_days;     //!< expiration time for lyrics cache files in days
     int cache_thumbs_keep_days;     //!< expiration time for thumbs cache files in days
     int cache_misc_keep_days;       //!< expiration time for misc cache files in days
+    int cache_http_keep_days;       //!< expiration time for HTTP cache files in days
     int http_port;                  //!< http port to listen
     int loglevel;                   //!< loglevel
     int ssl_port;                   //!< https port to listen