+++ src/main.c
@@ -280,6 +280,7 @@ static const struct t_subdirs_entry workdir_subdirs[] = {
  */
 static const struct t_subdirs_entry cachedir_subdirs[] = {
     {DIR_CACHE_COVER,         "Cover cache dir"},
+    {DIR_CACHE_HTTP,        "HTTP cache dir"},
     {DIR_CACHE_LYRICS,        "Lyrics cache dir"},
     {DIR_CACHE_MISC,          "Misc cache dir"},
     {DIR_CACHE_THUMBS, "Thumbs cache dir"},